# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "URL Shortener API",
      "version": "1.0.0",
      "description": "API for shortening URLs, redirecting to shortened URLs, and getting stats about shortened URLs."
    },
    "servers": [
      {
        "url": "/api"
      }
    ],
    "paths": {
      "/url/shorten": {
        "post": {
          "summary": "Shorten a URL",
          "description": "Accepts an original URL and an expiration date, and returns a shortened URL.",
          "operationId": "shortenUrl",
          "tags": ["URL"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "originalUrl": {
                      "type": "string",
                      "format": "uri",
                      "description": "The original URL to shorten."
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "The expiration date of the shortened URL."
                    }
                  },
                  "required": ["originalUrl", "expiresAt"]
                },
                "examples": {
                  "example1": {
                    "value": {
                      "originalUrl": "https://example.com",
                      "expiresAt": "2025-12-31T23:59:59Z"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Shortened URL successfully created.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "shortUrl": {
                        "type": "string",
                        "description": "The shortened URL."
                      }
                    }
                  },
                  "examples": {
                    "example1": {
                      "value": {
                        "shortUrl": "abc123"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad request due to invalid URL or date format."
            }
          }
        }
      },
      "/url/{shortUrl}": {
        "get": {
          "summary": "Redirect to shortened URL",
          "description": "Redirects to the original URL using the provided shortened URL.",
          "operationId": "redirectUrl",
          "tags": ["URL"],
          "parameters": [
            {
              "name": "shortUrl",
              "in": "path",
              "required": true,
              "description": "The shortened URL.",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "301": {
              "description": "Redirected to the original URL."
            },
            "400": {
              "description": "Invalid short URL."
            }
          }
        }
      },
      "/url/stats/{shortUrl}": {
        "get": {
          "summary": "Get stats for a shortened URL",
          "description": "Fetches statistics related to a shortened URL (e.g., number of redirects).",
          "operationId": "urlStats",
          "tags": ["URL"],
          "parameters": [
            {
              "name": "shortUrl",
              "in": "path",
              "required": true,
              "description": "The shortened URL.",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Stats retrieved successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "shortUrl": {
                        "type": "string",
                        "description": "The shortened URL."
                      },
                      "redirectCount": {
                        "type": "integer",
                        "description": "The number of times the shortened URL has been accessed."
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the shortened URL was created."
                      }
                    }
                  },
                  "examples": {
                    "example1": {
                      "value": {
                        "shortUrl": "abc123",
                        "redirectCount": 100,
                        "createdAt": "2025-02-13T10:00:00Z"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid short URL."
            }
          }
        }
      }
    }
  }
  